# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://localhost/open-api",
            "description": "API server"
        }
    ],
    "paths": {
        "/comments": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get list of blog comments",
                "operationId": "aca51a2bcfebde312eea5ac7ad4eb0c1",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filter",
                        "style": "deepObject",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "post_id": {
                                    "type": "integer"
                                },
                                "user_id": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "examples": {
                            "1": {
                                "summary": "1",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "tags": [
                    "file"
                ],
                "summary": "upload image",
                "operationId": "848b281d1a9e7bfea22f80d6f9dbb3ca",
                "requestBody": {
                    "description": "Upload images request body",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "file_url": {
                                            "description": "Url file",
                                            "format": "string",
                                            "default": "/images/examples.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Info"
                ],
                "operationId": "06eb65975fcc27d4b25c78658cc43cd4",
                "responses": {
                    "200": {
                        "description": "API version"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get list of blog posts",
                "operationId": "c50f64518c5d154fc54ca755f99a6a50",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "id",
                                    "-id",
                                    "title",
                                    "created_at"
                                ]
                            }
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filter",
                        "style": "deepObject",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "slug": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "offset",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "examples": {
                            "0": {
                                "summary": "0",
                                "value": "0"
                            }
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "examples": {
                            "10": {
                                "summary": "10",
                                "value": "10"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Store new post",
                "operationId": "516186c61bae40d1cc3e6fc16d7ceecf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "excerpt": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get post detail",
                "operationId": "ba0cf464d669dfe16bd554814e1be269",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "posts"
                ],
                "summary": "Update existing post",
                "operationId": "5bcd5ce3edb4ba418ba521fc92b44bf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "excerpt": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete existing post",
                "operationId": "5732f086ce06e69037fe8c729787f12f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/users/auth-token": {
            "post": {
                "tags": [
                    "Authorize"
                ],
                "summary": "User Login",
                "operationId": "d814a199546d08b1df6d5ed7c9053bec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "format": "string",
                                        "default": "schuster.montana@example.com"
                                    },
                                    "password": {
                                        "format": "string",
                                        "default": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "token",
                                            "format": "string",
                                            "default": "8|A3Id3nWYcxjGO2b30L0EknKT6999KRWme051zHpN"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CommentResource": {
                "properties": {
                    "id": {
                        "format": "int64",
                        "example": 1
                    },
                    "post_id": {
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "format": "int64",
                        "example": 1
                    },
                    "content": {
                        "format": "string",
                        "example": "Eum asperiores cumque inventore assumenda"
                    }
                },
                "type": "object"
            },
            "PostResource": {
                "properties": {
                    "id": {
                        "format": "int64",
                        "example": "1"
                    },
                    "category_id": {
                        "format": "int64",
                        "example": "1"
                    },
                    "title": {
                        "format": "string",
                        "example": "Example Post Name"
                    },
                    "slug": {
                        "format": "string",
                        "example": "post-slug-url"
                    },
                    "excerpt": {
                        "format": "string",
                        "example": "lorem lorem lorem lorem"
                    },
                    "content": {
                        "format": "string",
                        "example": "lorem ipsum lorem ipsumlorem ipsum lorem ipsum lorem ipsum"
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "1"
                            },
                            "name": {
                                "type": "string"
                            },
                            "default": {
                                "type": "boolean"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    }
}